\name{grafica_bonita}
\alias{grafica_bonita}
\title{Gráfica bonita}
\description{Genera una gráfica estilizada con leyenda, intervalos e indicador proyectado.}
\usage{
grafica_bonita(data, x, y, titulo, linea_vertical, mostrar_intervalo, fuente,
               etiqueta_x, etiqueta_y, mostrar_leyenda, nombre_estimado,
               nombre_superior, nombre_inferior)
}
\value{Un objeto \code{ggplot}}
\author{cpyp-ipdp}
